refactor: Remove unused file and Adjust site data folder structure
