+++ app/theme-providers.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { ThemeProvider } from 'next-themes'
-import siteMetadata from '@/data/siteMetadata'
+import siteMetadata from '@/data/site/siteMetadata'
 
 export function ThemeProviders({ children }: { children: React.ReactNode }) {
   return (
