+++ components/Footer.tsx
@@ -1,5 +1,5 @@
 import Link from './Link'
-import siteMetadata from '@/data/siteMetadata'
+import siteMetadata from '@/data/site/siteMetadata'
 import SocialIcon from '@/components/social-icons'
 
 export default function Footer() {
