+++ components/Header.tsx
@@ -1,6 +1,6 @@
-import siteMetadata from '@/data/siteMetadata'
-import headerNavLinks from '@/data/headerNavLinks'
-import Logo from '@/data/logo.svg'
+import siteMetadata from '@/data/site/siteMetadata'
+import headerNavLinks from '@/data/site/headerNavLinks'
+import Logo from '@/data/site/logo.svg'
 import Link from './Link'
 import MobileNav from './MobileNav'
 import ThemeSwitch from './ThemeSwitch'
