+++ data/blog/blog-getting-start.mdx
@@ -1,5 +1,5 @@
 ---
-title: '開始囉'
+title: '2023'
 date: '2023-12-16'
 lastmod: '2023-12-16'
 tags: ['Life']
