--- data/references-data.bib
@@ -1,33 +1,0 @@
-@article{Nash1950,
-  title={Equilibrium points in n-person games},
-  author={Nash, John},
-  journal={Proceedings of the national academy of sciences},
-  volume={36},
-  number={1},
-  pages={48--49},
-  year={1950},
-  publisher={USA}
-}
-
-@article{Nash1951,
-  title={Non-cooperative games},
-  author={Nash, John},
-  journal={Annals of mathematics},
-  pages={286--295},
-  year={1951},
-  publisher={JSTOR}
-}
-
-@Manual{Macfarlane2006,
-  url={https://pandoc.org/},
-  title={Pandoc: a universal document converter},
-  author={MacFarlane, John},
-  year={2006}
-}
-
-@book{Xie2016,
-  title={Bookdown: authoring books and technical documents with R markdown},
-  author={Xie, Yihui},
-  year={2016},
-  publisher={CRC Press}
-}
