+++ layouts/ListLayoutWithTags.tsx
@@ -8,7 +8,7 @@ import { CoreContent } from 'pliny/utils/contentlayer'
 import type { Blog } from 'contentlayer/generated'
 import Link from '@/components/Link'
 import Tag from '@/components/Tag'
-import siteMetadata from '@/data/siteMetadata'
+import siteMetadata from '@/data/site/siteMetadata'
 import tagData from 'app/tag-data.json'
 
 interface PaginationProps {
